Renamed Agent to Telegraph and moved it to msg package.
