+++ extensions/gdx-ai/src/com/badlogic/gdx/ai/msg/Telegraph.java
@@ -14,12 +14,11 @@
  * limitations under the License.
  ******************************************************************************/
 
-package com.badlogic.gdx.ai;
+package com.badlogic.gdx.ai.msg;
 
-import com.badlogic.gdx.ai.msg.Telegram;
 
 /** @author davebaol */
-public interface Agent {
+public interface Telegraph {
 
 	public boolean handleMessage (Telegram msg);
